requirements.txt
flask
flask-cors
PyJWT
gunicorn
python-dotenv
